changes to support Yeedu_2.9.0 version
